Camera as SDL_FRect: float maths, round only the final draw position; nearest-filter textures; L03 frame-cap/dt brought in line with the reference
Merges the L01-L03 camera decision forward onto L04's already-built
entity system. Render::camera becomes SDL_FRect instead of SDL_Rect --
SDL3's renderer is float-native, so the old int camera forced a
float -> int -> float round-trip on every move and truncated sub-pixel
motion. DrawTexture, DrawRectangle, DrawLine and DrawCircle now do
their camera-relative maths entirely in float and round only the final
destination once, last -- rounding only the camera isn't enough, since
entities at non-integer positions would still land between pixels.

Window::scale and Render's local scale become float to match, with
Input's mouse-scale read updated to the new type.

Textures::LoadSurface sets SDL_SCALEMODE_NEAREST right after creating
each texture, since SDL3 defaults to linear filtering, which blurs
pixel art and bleeds colour between adjacent tiles in a tileset.

Scene.h/.cpp: removed the cameraX/cameraY float-shadow members and the
(int) cast into render->camera -- they existed only to dodge the old
int camera's truncation; with camera.x/y themselves float, TODO 3's
comment and body are simplified to match (camSpeed * dt assigned
straight into render->camera.x/y). L04's own additions (the player
member, the L04: TODO 5 marker) are untouched.

Engine.cpp/.h: replaced this branch's own frame-cap/dt implementation
(int maxFrameDuration + delayRemainder carry, SDL_Delay) with the
reference's (float targetFrameRate, SDL_DelayPrecise), so Engine.cpp
matches L03_DeltaTime_Solution exactly apart from L04's own additions
(the EntityManager include/member and the L04: TODO 1 markers).

Reference: L03_DeltaTime_Solution.

+++ src/Input.cpp
@@ -112,7 +112,7 @@ bool Input::PreUpdate()
 
 		case SDL_EVENT_MOUSE_MOTION:
 		{
-			int scale = Engine::GetInstance().window->GetScale();
+			float scale = Engine::GetInstance().window->GetScale();
 			mouseMotionX = (int) (event.motion.xrel / scale);
 			mouseMotionY = (int)(event.motion.yrel / scale);
 			mouseX = (int)(event.motion.x / scale);
